Enabling Sign Comparison and warnings fixes

+++ Adjust/ADJActivityPackage.m
@@ -27,7 +27,7 @@
         NSArray * sortedKeys = [[self.parameters allKeys] sortedArrayUsingSelector:@selector(localizedStandardCompare:)];
         NSUInteger keyCount = [sortedKeys count];
         [builder appendFormat:@"Parameters:"];
-        for (int i = 0; i < keyCount; i++) {
+        for (NSUInteger i = 0; i < keyCount; i++) {
             NSString *key = (NSString*)[sortedKeys objectAtIndex:i];
             NSString *value = [self.parameters objectForKey:key];
             [builder appendFormat:@"\n\t\t%-22s %@", [key UTF8String], value];
